+++ CloudHarvestCoreTasks/tasks/__init__.py
@@ -7,6 +7,7 @@ from CloudHarvestCoreTasks.tasks.base import (
 # Tasks
 from CloudHarvestCoreTasks.tasks.dataset import DataSetTask
 from CloudHarvestCoreTasks.tasks.dummy import DummyTask
+from CloudHarvestCoreTasks.tasks.enqueue import EnqueueTask
 from CloudHarvestCoreTasks.tasks.error import ErrorTask
 from CloudHarvestCoreTasks.tasks.file import FileTask
 from CloudHarvestCoreTasks.tasks.harvest_update import HarvestUpdateTask